refactoring resources
