+++ src/__init__.py
@@ -1,5 +1,5 @@
 # flake8: noqa
 from .config import get_config
-from .user import User
+from src.resource.user import User
 from .models import SIR, SEIR, IModel
