+++ test/test_app.py
@@ -1,8 +1,9 @@
 import json
-import app
 from unittest import TestCase
 from werkzeug import exceptions
 from os import path
+
+from src import deprecated
 
 ABSOLUTE_PATH = path.dirname(path.abspath(__file__))
 
@@ -23,18 +24,18 @@ class TestApp(TestCase):
             self.expected_result_seir = json.load(json_file)
 
     def test_get_sir_data(self):
-        tmp = app.get_corona_data(param=self.param_sir)
+        tmp = deprecated.get_corona_data(param=self.param_sir)
         self.assertTrue(isinstance(tmp, str))
         self.assertEqual(json.loads(tmp), self.expected_result_sir)
 
     def test_get_seir_data(self):
-        tmp = app.get_corona_data(param=self.param_seir)
+        tmp = deprecated.get_corona_data(param=self.param_seir)
         self.assertTrue(isinstance(tmp, str))
         self.assertEqual(json.loads(tmp), self.expected_result_seir)
 
     def test_get_sir_data_pw(self):
         try:
-            app.get_corona_data(param={
+            deprecated.get_corona_data(param={
                 "username": "john",
                 "password": "yo"
             })
